market-chat-web/requirements.txt
# Streamlit web framework
streamlit>=1.28.0

# Data handling and analysis
pandas==2.2.3
numpy==2.1.3
requests==2.32.3

# Financial data APIs
yfinance==0.2.43

# Charting and visualization
plotly>=5.18.0

# Utilities
python-dotenv==1.0.1

# LLM Integration
groq>=0.33.0
